refactor: move functionality from base classes
